+++ src/QueryBuilder/QueryBuilder/QueryBuilderModal/DataTypeInput/DataTypeInput.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import { Pluggable } from '@folio/stripes/core'
 import { Select,
   TextField,
   MultiSelection,
@@ -37,6 +38,7 @@ export const DataTypeInput = ({
     OPERATORS.NOT_CONTAINS_ALL,
   ].includes(operator);
   const hasSourceOrValues = source || availableValues;
+  console.log(rest.fieldName, dataType)
 
   const { tenantTimezone: timezone } = useTenantTimezone();
 
@@ -112,8 +114,43 @@ export const DataTypeInput = ({
   };
 
   const stringTypeControls = () => {
+    const { fieldName } = rest;
+    const isPluginOrganizationRequired = fieldName === 'name' || fieldName === 'code';
     const isInRelatedWithOptions = isInRelatedOperator && hasSourceOrValues;
     const isEqualRelatedWithOptions = isEqualRelatedOperator && hasSourceOrValues;
+
+    if (isInRelatedWithOptions && isPluginOrganizationRequired) {
+      return (
+          <div className={className} data-testid="data-input-select-multi-stringType">
+            {multiSelectControl({testId: "data-input-select-multi-stringType"})}
+            <Pluggable
+                id={`${fieldName}-plugin`}
+                aria-haspopup="true"
+                dataKey="organization"
+                searchButtonStyle="link"
+                searchLabel={<FormattedMessage id="stripes-acq-components.filter.organization.lookup"/>}
+                selectVendor={(selectedItems)=> {
+                  const normalizedItems = selectedItems.map(item => {
+                    if(fieldName === 'code') {
+                      return ({
+                        id: item.id,
+                        label: item.code,
+                      })
+                    } return ({
+                    id: item.id,
+                    label: item.name,
+                  })});
+                  onChange(normalizedItems, index, COLUMN_KEYS.VALUE)}
+                }
+                type="find-organization"
+                usePortal={true}
+                isMultiSelect
+            >
+              <FormattedMessage id="stripes-acq-components.filter.organization.lookupNoSupport"/>
+            </Pluggable>
+          </div>
+      );
+    }
 
     if (isInRelatedWithOptions) {
       return (
@@ -123,32 +160,52 @@ export const DataTypeInput = ({
       );
     }
 
+    if (isEqualRelatedWithOptions && isPluginOrganizationRequired) {
+      return (
+          <div className={className}>
+            {selectControl({testId: 'data-input-select-single-stringType'})}
+            <Pluggable
+                id={`${fieldName}-plugin`}
+                aria-haspopup="true"
+                dataKey="organization"
+                searchButtonStyle="link"
+                searchLabel={<FormattedMessage id="stripes-acq-components.filter.organization.lookup"/>}
+                selectVendor={(e)=> onChange(e.id, index, COLUMN_KEYS.VALUE)}
+                type="find-organization"
+                usePortal={true}
+            >
+              <FormattedMessage id="stripes-acq-components.filter.organization.lookupNoSupport"/>
+            </Pluggable>
+          </div>
+      )
+    }
+
     if (isEqualRelatedWithOptions) {
       return (
-        <div className={className}>
-          {selectControl({ testId: 'data-input-select-single-stringType' })}
-        </div>
+          <div className={className}>
+            {selectControl({testId: 'data-input-select-single-stringType'})}
+          </div>
       );
     }
 
-    return textControl({ testId: 'data-input-text-stringType' });
+    return textControl({testId: 'data-input-text-stringType'});
   };
 
   const numericTypeControls = () => {
     return hasSourceOrValues
-      ? selectControl({ testId: 'data-input-select-numeric' })
-      : textControl({ type: 'number', textClass: css.NumberInput });
+        ? selectControl({testId: 'data-input-select-numeric'})
+        : textControl({type: 'number', textClass: css.NumberInput});
   };
 
   const booleanTypeControls = () => (
-    <div className={className}>
-      {selectControl({ testId: 'data-input-select-boolType' })}
-    </div>
+      <div className={className}>
+        {selectControl({testId: 'data-input-select-boolType'})}
+      </div>
   );
 
   const openUUIDTypeControls = () => {
     return isInRelatedOperator ? (
-      <>
+        <>
         {textAreaControl()}
         <FormattedMessage id="ui-plugin-query-builder.control.info.separateValues" />
       </>